Refactor import statements and fix class names
