--- agogos/transforming_system.py
@@ -1,65 +1,0 @@
-from typing import Any
-from agogos._core._system import _System
-from agogos.transformer import Transformer
-
-
-class TransformingSystem(_System):
-    """A system that transforms the input data.
-
-    ### Parameters:
-    - steps (list[Transformer | TransformingSystem | ParallelTransformingSystem]): The steps in the system.
-
-    Implements the following methods:
-    ```python
-    def transform(self, x: Any, transform_args: dict[str, Any] = {}) -> Any: # Transform the input data.
-
-    def get_hash(self) -> str: # Get the hash of the system.
-    ```
-
-    Usage:
-    ```python
-    from agogos.transforming_system import TransformingSystem
-
-    transformer_1 = CustomTransformer()
-    transformer_2 = CustomTransformer()
-
-    transforming_system = TransformingSystem(steps=[transformer_1, transformer_2])
-    transformed_data = transforming_system.transform(data)
-    predictions = transforming_system.predict(data)
-    ```
-    """
-
-    def __post_init__(self) -> None:
-        """Post init method for the TransformingSystem class."""
-
-        # Assert all steps are a subclass of Transformer
-        for step in self.steps:
-            assert issubclass(step.__class__, Transformer) or issubclass(
-                step.__class__, TransformingSystem
-            ), f"{step} is not a subclass of Transformer"
-
-        super().__post_init__()
-
-    def transform(self, x: Any, transform_args: dict[str, Any] = {}) -> Any:
-        """Transform the input data.
-
-        :param x: The input data.
-        :return: The transformed data.
-        """
-
-        # Loop through each step and call the transform method
-        for step in self.steps:
-            step_name = step.__class__.__name__
-            step_transform_args = (
-                transform_args[step_name]
-                if transform_args and step_name in transform_args
-                else {}
-            )
-            if isinstance(step, Transformer):
-                x = step.transform(x, **step_transform_args)
-            elif isinstance(step, TransformingSystem):
-                x = step.transform(x, step_transform_args)
-            else:
-                raise TypeError(f"{step} is not a subclass of Transformer")
-
-        return x
