+++ tests/test_transforming_system.py
@@ -1,7 +1,6 @@
 import numpy as np
 import pytest
-from agogos.transformer import Transformer
-from agogos.transforming_system import TransformingSystem
+from agogos.transforming import Transformer, TransformingSystem
 
 
 class TestTransformingSystem:
